1.	Created new RpmVisualizer.groovy class that extends Visualizer.groovy. Ditto for  RpmVisualizerInfo.groovy and RpmVisualizerRelInfo.groovy. 2.	Changed NCubeController to direct to Visualizer for non-rpm cubes and to RpmVisualizer for rpm class cubes. This code can be refactored when we move the visualizer out to a separate project. 3.	Rpm class cubes: the visualization works as before. 4.	Non-rpm cubes: the visualization will display the top level node only for now. Added toast that this is under construction.
