+++ src/main/groovy/com/cedarsoftware/util/Visualizer.groovy
@@ -10,23 +10,19 @@ import groovy.transform.CompileStatic
 import static com.cedarsoftware.util.VisualizerConstants.*
 
 /**
- * Provides the information used to visualize rpm cubes associated with a given rpm cube.
+ * Provides the information used to visualize n-cubes.
  */
 
-// TODO: This code needs to be moved out of NCE and pulled-in via Grapes.
 @CompileStatic
 class Visualizer
 {
-	private ApplicationID appId
-	private VisualizerHelper helper = new VisualizerHelper()
-	private Set<String> messages = []
-	private Set<String> visited = []
-	private String defaultScopeEffectiveVersion
-	private String defaultScopeDate
-	private Deque<VisualizerRelInfo> stack = new ArrayDeque<>()
+	protected ApplicationID appId
+	protected Set<String> messages = []
+	protected Set<String> visited = []
+	protected Deque<VisualizerRelInfo> stack = new ArrayDeque<>()
 
 	/**
-	 * Provides the information used to visualize rpm cubes associated with a given rpm cube.
+	 * Provides the information used to visualize n-cubes.
 	 *
 	 * @param applicationID
 	 * @param options - a map containing:
@@ -38,10 +34,8 @@ class Visualizer
 	Map<String, Object> buildGraph(ApplicationID applicationID, Map options)
 	{
 		appId = applicationID
-		defaultScopeEffectiveVersion = applicationID.version.replace('.', '-')
-		defaultScopeDate = DATE_TIME_FORMAT.format(new Date())
 		String startCubeName = options.startCubeName as String
-		VisualizerInfo visInfo = options.visInfo as VisualizerInfo ?: new VisualizerInfo(appId, options)
+		VisualizerInfo visInfo = getVisualizerInfo(options)
 
 		if (!isValidStartCube(startCubeName))
 		{
@@ -53,93 +47,51 @@ class Visualizer
 			return [status: STATUS_MISSING_START_SCOPE, visInfo: visInfo, message: messages.join(DOUBLE_BREAK)]
 		}
 
-		getRpmVisualization(visInfo, startCubeName)
+		getVisualization(visInfo, startCubeName)
 
 		String message = messages.empty ? null : messages.join(DOUBLE_BREAK)
 		return [status: STATUS_SUCCESS, visInfo: visInfo, message: message]
 	}
 
-	/**
-	 * Loads all traits available for a given class.
-	 *
-	 * @param applicationID
-	 * @param options (map containing:
-	 *            Map node, representing a class,
-	 *            VisualizerInfo, information about the visualization
-	 * @return node
-     */
-	Map getTraits(ApplicationID applicationID, Map options)
+	protected VisualizerInfo getVisualizerInfo(Map options)
 	{
-		appId = applicationID
-
-		VisualizerInfo visInfo = options.visInfo as VisualizerInfo
-		Map node = options.node as Map
-		VisualizerRelInfo relInfo = new VisualizerRelInfo(appId, visInfo.allGroupsKeys, node)
-
-		getTraitMaps(visInfo, relInfo)
-		node.details = relInfo.details
-		visInfo.nodes = [node]
-
-		String message = messages.empty ? null : messages.join(DOUBLE_BREAK)
-		return [status: STATUS_SUCCESS, visInfo: visInfo, message: message]
+		VisualizerInfo visInfo = options.visInfo as VisualizerInfo ?: new VisualizerInfo(appId, options)
+		visInfo.scope = visInfo.scope as CaseInsensitiveMap ?: new CaseInsensitiveMap<>()
+		return visInfo
 	}
 
-	private void getRpmVisualization(VisualizerInfo visInfo, String startCubeName)
+	protected void getVisualization(VisualizerInfo visInfo, String startCubeName)
 	{
 		visInfo.maxLevel = 1
 		visInfo.nodeCount = 1
 		visInfo.availableGroupsAllLevels = [] as Set
-
-		VisualizerRelInfo relInfo = new VisualizerRelInfo()
-		relInfo.targetCube = NCubeManager.getCube(appId, startCubeName)
-		relInfo.scope = visInfo.scope
-		relInfo.targetLevel = 1
-		relInfo.targetId = 1
-		relInfo.typesToAdd = visInfo.getTypesToAdd(relInfo.targetCube.name)
+		VisualizerRelInfo relInfo = getVisualizerRelInfo(visInfo, startCubeName)
 		stack.push(relInfo)
 
 		while (!stack.empty)
 		{
 			processCube(visInfo, stack.pop())
 		}
-
-		/* TODO: Not needed currently, but will revisit
-		addSets(visInfo.availableScopeKeys, visInfo.requiredScopeKeys.values() as Set)
-		addSets(visInfo.availableScopeKeys, visInfo.optionalScopeKeys.values() as Set)
-		*/
 	}
 
-	/*
-	TODO: Not needed currently, but will revisit
-	private static Set<Set> addSets(Set<String> set, Set<Set> sets)
+	protected VisualizerRelInfo getVisualizerRelInfo(VisualizerInfo visInfo, String startCubeName)
 	{
-		sets.each {
-			set.addAll(it)
-		}
-		return sets
-	}*/
-
-	private void processCube(VisualizerInfo visInfo, VisualizerRelInfo relInfo)
-	{
-		if (relInfo.targetCube.name.startsWith(RPM_CLASS))
-		{
-			processClassCube(visInfo, relInfo)
-		}
-		else
-		{
-			processEnumCube(visInfo, relInfo)
-		}
+		VisualizerRelInfo relInfo = new VisualizerRelInfo()
+		relInfo.targetCube = NCubeManager.getCube(appId, startCubeName)
+		relInfo.scope = visInfo.scope
+		relInfo.targetLevel = 1
+		relInfo.targetId = 1
+		relInfo.targetScope = [:]
+		return relInfo
 	}
 
-	private void processClassCube(VisualizerInfo visInfo, VisualizerRelInfo relInfo)
+	protected void processCube(VisualizerInfo visInfo, VisualizerRelInfo relInfo)
 	{
+		messages << "*** UNDER CONSTRUCTION *** ${DOUBLE_BREAK} Full visualization of non-rpm class cubes is not yet available.".toString()
+
 		String targetCubeName = relInfo.targetCube.name
 
-		boolean hasFields = canLoadTargetAsRpmClass(relInfo)
-		if (hasFields)
-		{
-			hasFields = getTraitMaps(visInfo, relInfo)
-		}
+		relInfo.addRequiredAndOptionalScopeKeys(visInfo)
 
 		if (relInfo.sourceCube)
 		{
@@ -154,122 +106,31 @@ class Visualizer
 		visInfo.nodes << relInfo.createNode(visInfo.allGroupsKeys, visInfo.groupSuffix)
 		visInfo.availableGroupsAllLevels << relInfo.group
 
-		if (hasFields)
+
+		NCube nextTargetCube
+		String nextTargetCubeName = ""
+
+		if (nextTargetCube)
 		{
-			relInfo.targetTraitMaps.each { String targetFieldName, Map targetTraits ->
-				if (CLASS_TRAITS != targetFieldName)
-				{
-					String targetFieldRpmType = targetTraits[R_RPM_TYPE]
-
-					if (!helper.isPrimitive(targetFieldRpmType))
-					{
-						NCube nextTargetCube
-						String nextTargetCubeName = ""
-						if (targetTraits.containsKey(V_ENUM))
-						{
-							nextTargetCubeName = RPM_ENUM_DOT + targetTraits[V_ENUM]
-							nextTargetCube = NCubeManager.getCube(appId, nextTargetCubeName)
-						}
-						else if (targetFieldRpmType)
-						{
-							nextTargetCubeName = RPM_CLASS_DOT + targetFieldRpmType
-							nextTargetCube = NCubeManager.getCube(appId, nextTargetCubeName)
-						}
-
-						if (nextTargetCube)
-						{
-							addToStack(visInfo, relInfo, nextTargetCube, targetFieldRpmType, targetFieldName)
-						}
-						else
-						{
-							messages << "No cube exists with name of ${nextTargetCubeName}. Cube not included in the visualization.".toString()
-						}
-					}
-				}
-			}
+			addToStack(visInfo, relInfo, nextTargetCube)
 		}
 	}
 
-	private void processEnumCube(VisualizerInfo visInfo, VisualizerRelInfo relInfo)
+	protected VisualizerRelInfo addToStack(VisualizerInfo visInfo, VisualizerRelInfo relInfo, NCube nextTargetCube)
 	{
-		relInfo.group = UNSPECIFIED
-		String targetCubeName = relInfo.targetCube.name
-		String sourceFieldRpmType = relInfo.sourceFieldRpmType
 
-		if (!targetCubeName.startsWith(RPM_ENUM))
-		{
-			throw new IllegalStateException("Cube is not an rpm.enum cube: ${targetCubeName}.")
-		}
-
-		if (relInfo.sourceCube && (!sourceFieldRpmType || helper.isPrimitive(sourceFieldRpmType)))
-		{
-			return
-		}
-
-		boolean hasFields = getTraitMaps(visInfo, relInfo)
-
-		if (hasFields)
-		{
-			relInfo.targetTraitMaps.each { String targetFieldName, Map targetTraits ->
-				if (CLASS_TRAITS != targetFieldName)
-				{
-					try
-					{
-						String nextTargetCubeName = relInfo.getNextTargetCubeName(targetFieldName)
-
-						if (nextTargetCubeName)
-						{
-							NCube nextTargetCube = NCubeManager.getCube(appId, nextTargetCubeName)
-							if (nextTargetCube)
-							{
-								addToStack(visInfo, relInfo, nextTargetCube, relInfo.sourceFieldRpmType, targetFieldName)
-
-								if (relInfo.group == UNSPECIFIED)
-								{
-									relInfo.setGroupName(visInfo.allGroupsKeys, nextTargetCubeName)
-								}
-							}
-							else
-							{
-								messages << "No cube exists with name of ${nextTargetCubeName}. Cube not included in the visualization.".toString()
-							}
-						}
-					}
-					catch (Exception e)
-					{
-						throw new IllegalStateException("Error processing the cube for enum field ${targetFieldName} in enum ${targetCubeName}.", e)
-					}
-				}
-			}
-		}
-
-		visInfo.edges << relInfo.createEdge(visInfo.edges.size())
-
-		if (!visited.add(targetCubeName + relInfo.scope.toString()))
-		{
-			return
-		}
-
-		visInfo.nodes << relInfo.createNode(visInfo.allGroupsKeys, visInfo.groupSuffix)
-		visInfo.availableGroupsAllLevels << relInfo.group
-	}
-
-	private void addToStack(VisualizerInfo visInfo, VisualizerRelInfo relInfo, NCube nextTargetCube, String rpmType, String targetFieldName)
-	{
+		VisualizerRelInfo nextRelInfo
 		try
 		{
 			NCube nextSourceCube = relInfo.targetCube
 
-			VisualizerRelInfo nextRelInfo = new VisualizerRelInfo()
+			nextRelInfo = visualizerRelInfo
 			nextRelInfo.targetCube = nextTargetCube
-			nextRelInfo.scope = getScopeRelativeToSource(nextTargetCube, rpmType, targetFieldName, relInfo.scope)
+			nextRelInfo.scope = [:]  //TODO
 			nextRelInfo.sourceCube = nextSourceCube
 			nextRelInfo.sourceScope = relInfo.targetScope
-			nextRelInfo.sourceFieldName = targetFieldName
-			nextRelInfo.sourceFieldRpmType = rpmType
-			nextRelInfo.sourceTraitMaps = relInfo.targetTraitMaps
+			nextRelInfo.sourceFieldName = '' //TODO
 			nextRelInfo.sourceId = relInfo.targetId
-			nextRelInfo.typesToAdd = visInfo.getTypesToAdd(nextTargetCube.name)
 
 			long nextTargetTargetLevel = relInfo.targetLevel + 1
 			nextRelInfo.targetLevel = nextTargetTargetLevel
@@ -283,172 +144,27 @@ class Visualizer
 		}
 		catch (Exception e)
 		{
-			throw new IllegalStateException("Error processing the class for field ${relInfo.sourceFieldName} in class ${nextTargetCube.name}.", e)
+			throw new IllegalStateException("Error processing ${relInfo.sourceFieldName} in cube ${nextTargetCube.name}.", e)
 		}
+		return nextRelInfo
 	}
 
-	private boolean canLoadTargetAsRpmClass(VisualizerRelInfo relInfo)
+	protected VisualizerRelInfo getVisualizerRelInfo()
 	{
-		//When the source cube points directly to the target cube (source cube and target cube are both rpm.class),
-		//check if the source field name matches up with the scoped name of the target. If not, the target cube cannot be
-		//loaded as an rpm.class.
-		NCube sourceCube = relInfo.sourceCube
-		NCube targetCube = relInfo.targetCube
+		return new VisualizerRelInfo()
+	}
 
-		if (sourceCube && sourceCube.name.startsWith(RPM_CLASS_DOT) && targetCube.name.startsWith(RPM_CLASS_DOT) &&
-				targetCube.getAxis(AXIS_TRAIT).findColumn(R_SCOPED_NAME))
-		{
-			String type = relInfo.sourceFieldRpmType
-			NCube classTraitsCube = NCubeManager.getCube(appId, RPM_SCOPE_CLASS_DOT + type + DOT_CLASS_TRAITS)
-			String sourceFieldName = relInfo.sourceFieldName
-			if (!classTraitsCube.getAxis(type).findColumn(sourceFieldName))
-			{
-				relInfo.targetTraitMaps = [(CLASS_TRAITS): [(R_SCOPED_NAME): UNABLE_TO_LOAD + relInfo.sourceFieldName]] as Map
-				String msg = getLoadTargetAsRpmClassMessage(relInfo, type)
-				relInfo.notes << msg
-				relInfo.hasFields = false
-				return false
-			}
-		}
+	protected boolean isValidStartCube(String cubeName)
+	{
 		return true
 	}
 
-
-	/**
-	 * Sets the basic scope required to load a target class based on scoped source class, source field name, target class name, and current scope.
-	 * Retains all other scope.
-	 * If the source class is not a scoped class, returns the scope unchanged.
-	 *
-	 * @param targetCube String target cube
-	 * @param targetTraitsMaps Map fields and traits of the target class
-	 * @param sourceTraitsMap Map fields and traits of the source class
-	 * @param sourceFieldName String source field name
-	 * @param scope Map<String, Object> scope
-	 *
-	 * @return Map new scope
-	 *
-	 */
-	private static Map<String, Object> getScopeRelativeToSource(NCube targetCube, String sourceFieldRpmType, String targetFieldName, Map scope)
+	protected boolean hasMissingMinimumScope(VisualizerInfo visInfo, String startCubeName)
 	{
-		Map<String, Object> newScope = new CaseInsensitiveMap<>(scope)
-
-		if (targetCube.name.startsWith(RPM_ENUM))
-		{
-			newScope[SOURCE_FIELD_NAME] = targetFieldName
-		}
-		else if (targetCube.getAxis(AXIS_TRAIT).findColumn(R_SCOPED_NAME))
-		{
-			String newScopeKey = sourceFieldRpmType
-			String oldValue = scope[newScopeKey]
-			if (oldValue)
-			{
-				newScope[SOURCE_SCOPE_KEY_PREFIX + sourceFieldRpmType] = oldValue
-			}
-			newScope[newScopeKey] = targetFieldName
-		}
-		return newScope
+		return false
 	}
 
-	private Map<String, Object> getDefaultScope(String type)
-	{
-		Map<String, Object> scope = new CaseInsensitiveMap<>()
-		if (type)
-		{
-			scope[type] = DEFAULT_SCOPE_VALUE
-			scope[POLICY_CONTROL_DATE] = defaultScopeDate
-			scope[QUOTE_DATE] = defaultScopeDate
-		}
-		scope[EFFECTIVE_VERSION] = defaultScopeEffectiveVersion
-		return scope
-	}
-
-	private boolean isValidStartCube(String cubeName)
-	{
-		if (!cubeName.startsWith(RPM_CLASS_DOT))
-		{
-			messages << "Starting cube for visualization must begin with 'rpm.class', n-cube ${cubeName} does not.".toString()
-			return false
-		}
-
-		NCube cube = NCubeManager.getCube(appId, cubeName)
-		if (!cube.getAxis(AXIS_FIELD) || !cube.getAxis(AXIS_TRAIT) )
-		{
-			messages << "Cube ${cubeName} is not a valid rpm class since it does not have both a field axis and a traits axis.".toString()
-			return false
-		}
-		return true
-	}
-
-	private boolean hasMissingMinimumScope(VisualizerInfo visInfo, String startCubeName)
-	{
-		boolean hasMissingScope = false
-		Map<String, Object> scope = visInfo.scope
-		String messageSuffixScopeKey = "Its default value may be changed as desired."
-
-		if (NCubeManager.getCube(appId, startCubeName).getAxis(AXIS_TRAIT).findColumn(R_SCOPED_NAME))
-		{
-			String type = getTypeFromCubeName(startCubeName)
-			String messageSuffixTypeScopeKey = "${DOUBLE_BREAK}Please replace ${DEFAULT_SCOPE_VALUE} for ${type} with an actual scope value."
-			String messageScopeValues = getAvailableScopeValuesMessage(visInfo, startCubeName, type)
-			String messageSuffix = 'The other default scope values may also be changed as desired.'
-			if (scope)
-			{
-				hasMissingScope = visInfo.addMissingMinimumScope(type, DEFAULT_SCOPE_VALUE, "${messageSuffixTypeScopeKey}${messageScopeValues}", messages) ?: hasMissingScope
-				hasMissingScope = visInfo.addMissingMinimumScope(POLICY_CONTROL_DATE, defaultScopeDate, messageSuffixScopeKey, messages) ?: hasMissingScope
-				hasMissingScope = visInfo.addMissingMinimumScope(QUOTE_DATE, defaultScopeDate, messageSuffixScopeKey, messages) ?: hasMissingScope
-				hasMissingScope = visInfo.addMissingMinimumScope(EFFECTIVE_VERSION, defaultScopeEffectiveVersion, messageSuffixScopeKey, messages) ?: hasMissingScope
-			}
-			else
-			{
-				hasMissingScope = true
-				Map<String, Object> defaultScope = getDefaultScope(type)
-				visInfo.scope = defaultScope
-				String msg = getMissingMinimumScopeMessage(defaultScope, messageScopeValues, messageSuffixTypeScopeKey, messageSuffix)
-				messages << msg
-			}
-		}
-		else{
-			if (scope)
-			{
-				hasMissingScope = visInfo.addMissingMinimumScope(EFFECTIVE_VERSION, defaultScopeEffectiveVersion, messageSuffixScopeKey, messages) ?: hasMissingScope
-			}
-			else
-			{
-				hasMissingScope = false
-				visInfo.scope = getDefaultScope(null)
-			}
-		}
-		return hasMissingScope
-	}
-
-	private boolean getTraitMaps(VisualizerInfo visInfo, VisualizerRelInfo relInfo)
-	{
-		try
-		{
-            relInfo.getTraitMaps(helper, appId, visInfo)
-			relInfo.hasFields = true
-		}
-		catch (Exception e)
-		{
-			relInfo.hasFields = false
-			Throwable t = getDeepestException(e)
-			if (t instanceof InvalidCoordinateException)
-			{
-				handleInvalidCoordinateException(t as InvalidCoordinateException, visInfo, relInfo)
-			}
-			else if (t instanceof CoordinateNotFoundException)
-			{
-				handleCoordinateNotFoundException(t as CoordinateNotFoundException, visInfo, relInfo)
-			}
-			else
-			{
-				handleException(t, relInfo)
-			}
-		}
-		return relInfo.hasFields
-	}
-
-	private void handleCoordinateNotFoundException(CoordinateNotFoundException e, VisualizerInfo visInfo, VisualizerRelInfo relInfo)
+	protected String handleCoordinateNotFoundException(CoordinateNotFoundException e, VisualizerInfo visInfo, VisualizerRelInfo relInfo)
 	{
 		String cubeName = e.cubeName
 		String axisName = e.axisName
@@ -460,17 +176,17 @@ class Visualizer
 			String msg = getCoordinateNotFoundMessage(visInfo, relInfo, axisName, value, effectiveName, cubeName)
 			relInfo.notes << msg
 			messages << msg
-			relInfo.targetTraitMaps = [(CLASS_TRAITS): [(R_SCOPED_NAME): SCOPE_VALUE_NOT_FOUND + effectiveName]] as Map
+			return SCOPE_VALUE_NOT_FOUND
 		}
 		else
 		{
-			handleException(e as Exception, relInfo)
+			return handleException(e as Exception, relInfo)
 		}
 	}
 
-	private void handleInvalidCoordinateException(InvalidCoordinateException e, VisualizerInfo visInfo, VisualizerRelInfo relInfo)
+	protected String handleInvalidCoordinateException(InvalidCoordinateException e, VisualizerInfo visInfo, VisualizerRelInfo relInfo, Set mandatoryScopeKeys)
 	{
-		Set<String> missingScope = findMissingScope(relInfo.scope, e.requiredKeys)
+		Set<String> missingScope = findMissingScope(relInfo.scope, e.requiredKeys, mandatoryScopeKeys)
 
 		if (missingScope)
 		{
@@ -480,10 +196,10 @@ class Visualizer
 			}
 			visInfo.scope = expandedScope
 			String effectiveName = relInfo.effectiveNameByCubeName
-			relInfo.targetTraitMaps = [(CLASS_TRAITS): [(R_SCOPED_NAME): MISSING_SCOPE + effectiveName]] as Map
 			String msg = getInvalidCoordinateExceptionMessage(visInfo, relInfo, missingScope, effectiveName, e.cubeName)
 			relInfo.notes << msg
 			messages << msg
+			return MISSING_SCOPE
 		}
 		else
 		{
@@ -491,24 +207,24 @@ class Visualizer
 		}
 	}
 
-	private static Set<String> findMissingScope(Map<String, Object> scope, Set<String> requiredKeys)
+	protected static Set<String> findMissingScope(Map<String, Object> scope, Set<String> requiredKeys, Set mandatoryScopeKeys)
 	{
 		return requiredKeys.findAll { String key ->
-            !MANDATORY_RPM_SCOPE_KEYS.contains(key) && (scope == null || !scope.containsKey(key))
+            !mandatoryScopeKeys.contains(key) && (scope == null || !scope.containsKey(key))
         }
 	}
 
-	private void handleException(Throwable e, VisualizerRelInfo relInfo)
+	protected String handleException(Throwable e, VisualizerRelInfo relInfo)
 	{
 		Throwable t = getDeepestException(e)
 		String effectiveName = relInfo.effectiveNameByCubeName
 		String msg = getExceptionMessage(relInfo, effectiveName, e, t)
 		relInfo.notes << msg
 		messages << msg
-		relInfo.targetTraitMaps = [(CLASS_TRAITS): [(R_SCOPED_NAME): UNABLE_TO_LOAD + effectiveName]] as Map
+		return UNABLE_TO_LOAD
 	}
 
-	private static Throwable getDeepestException(Throwable e)
+	protected static Throwable getDeepestException(Throwable e)
 	{
 		while (e.cause != null)
 		{
@@ -517,12 +233,7 @@ class Visualizer
 		return e
 	}
 
-	private static String getTypeFromCubeName(String cubeName)
-	{
-		return (cubeName - RPM_CLASS_DOT)
-	}
-
-	private static String getAvailableScopeValuesMessage(VisualizerInfo visInfo, String cubeName, String key)
+	protected static String getAvailableScopeValuesMessage(VisualizerInfo visInfo, String cubeName, String key)
 	{
 		Set<Object> scopeValues = visInfo.availableScopeValues[key] ?: visInfo.loadAvailableScopeValues(cubeName, key)
 		if (scopeValues) {
@@ -538,7 +249,7 @@ class Visualizer
 		return ''
 	}
 
-	private static String getMissingMinimumScopeMessage(Map<String, Object> scope, String messageScopeValues, String messageSuffixType, String messageSuffix )
+	protected static String getMissingMinimumScopeMessage(Map<String, Object> scope, String messageScopeValues, String messageSuffixType, String messageSuffix )
 	{
 		"""\
 The scope for the following scope keys was added since it was required: \
@@ -547,59 +258,31 @@ ${messageSuffixType} ${messageSuffix} \
 ${BREAK}${messageScopeValues}"""
 	}
 
-	private static String getExceptionMessage(VisualizerRelInfo relInfo, String effectiveName, Throwable e, Throwable t)
+	protected static String getExceptionMessage(VisualizerRelInfo relInfo, String effectiveName, Throwable e, Throwable t)
 	{
 		String cubeDisplayName = relInfo.getCubeDisplayName(relInfo.targetCube.name)
 		"""\
-An exception was thrown while loading fields and traits for ${cubeDisplayName}${relInfo.sourceMessage} for ${effectiveName}. \
+An exception was thrown while loading for ${cubeDisplayName}${relInfo.sourceMessage} for ${effectiveName}. \
 ${DOUBLE_BREAK}<b>Message:</b> ${DOUBLE_BREAK}${e.message}${DOUBLE_BREAK}<b>Root cause: </b>\
 ${DOUBLE_BREAK}${t.toString()}${DOUBLE_BREAK}<b>Stack trace: </b>${DOUBLE_BREAK}${t.stackTrace.toString()}"""
 	}
 
-	private static String getCoordinateNotFoundMessage(VisualizerInfo visInfo, VisualizerRelInfo relInfo, String key, Object value, String effectiveName, String cubeName)
+	protected String getCoordinateNotFoundMessage(VisualizerInfo visInfo, VisualizerRelInfo relInfo, String key, Object value, String effectiveName, String cubeName)
 	{
 		StringBuilder message = new StringBuilder()
 		String messageScopeValues = getAvailableScopeValuesMessage(visInfo, cubeName, key)
-		if (cubeName.startsWith(RPM_CLASS_DOT) || cubeName.startsWith(RPM_ENUM_DOT))
-		{
-			String cubeDisplayName = relInfo.getCubeDisplayName(cubeName)
-			message.append("The scope value ${value} for scope key ${key} cannot be found on axis ${key} in ${cubeDisplayName}${relInfo.sourceMessage} for ${effectiveName}.")
-		}
-		else
-		{
-			message.append("The scope value ${value} for scope key ${key} cannot be found on axis ${key} in cube ${cubeName} for ${effectiveName}.")
-		}
+		message.append("The scope value ${value} for scope key ${key} cannot be found on axis ${key} in cube ${cubeName} for ${effectiveName}.")
 		message.append("${DOUBLE_BREAK} Please supply a different value for ${key}.${BREAK}${messageScopeValues}")
 	}
 
-	private static String getInvalidCoordinateExceptionMessage(VisualizerInfo visInfo, VisualizerRelInfo relInfo, Set<String> missingScope, String effectiveName, String cubeName)
+	protected String getInvalidCoordinateExceptionMessage(VisualizerInfo visInfo, VisualizerRelInfo relInfo, Set<String> missingScope, String effectiveName, String cubeName)
 	{
 		StringBuilder message = new StringBuilder()
-
-		if (cubeName.startsWith(RPM_CLASS_DOT) || cubeName.startsWith(RPM_ENUM_DOT))
-		{
-			String cubeDisplayName = relInfo.getCubeDisplayName(cubeName)
-			message.append("Additional scope is required to load ${effectiveName} of type ${cubeDisplayName}${relInfo.sourceMessage}.")
-		}
-		else
-		{
-			message.append("Additional scope is required to load cube ${cubeName} for ${effectiveName}${relInfo.sourceMessage}.")
-		}
+		message.append("Additional scope is required to load cube ${cubeName} for ${effectiveName}${relInfo.sourceMessage}.")
 		message.append("${DOUBLE_BREAK} Please add scope value(s) for the following scope key(s): ${missingScope.join(COMMA_SPACE)}.${BREAK}")
 		missingScope.each{ String key ->
 			message.append(getAvailableScopeValuesMessage(visInfo, cubeName, key))
 		}
 		return message.toString()
-	}
-
-	private static String getLoadTargetAsRpmClassMessage(VisualizerRelInfo relInfo, String type) {
-
-		String sourceCubeDisplayName = relInfo.getCubeDisplayName(relInfo.sourceCube.name)
-		String targetCubeDisplayName = relInfo.getCubeDisplayName(relInfo.targetCube.name)
-
-		"""\
-${sourceCubeDisplayName} points directly to ${targetCubeDisplayName} via field ${relInfo.sourceFieldName}, but \
-there is no ${type.toLowerCase()} named ${relInfo.sourceFieldName} on ${type}.  ${DOUBLE_BREAK}Therefore \
-it cannot be loaded in the visualization."""
 	}
 }
