+++ src/main/groovy/com/cedarsoftware/util/VisualizerRelInfo.groovy
@@ -3,8 +3,6 @@ package com.cedarsoftware.util
 import com.cedarsoftware.ncube.ApplicationID
 import com.cedarsoftware.ncube.NCube
 import com.cedarsoftware.ncube.NCubeManager
-import com.cedarsoftware.ncube.RuleInfo
-import com.google.common.base.Splitter
 import groovy.transform.CompileStatic
 
 import static com.cedarsoftware.util.VisualizerConstants.*
@@ -16,8 +14,6 @@ import static com.cedarsoftware.util.VisualizerConstants.*
 @CompileStatic
 class VisualizerRelInfo
 {
-	boolean hasFields = false
-	boolean loadTraits = false
 	Set<String> notes = []
 	Map<String, Object> scope
 	String group
@@ -25,16 +21,12 @@ class VisualizerRelInfo
 	long targetId
     NCube targetCube
 	Map<String, Object> targetScope
-	Map<String, Map<String, Object>> targetTraitMaps
 	long targetLevel
 
 	long sourceId
 	NCube sourceCube
 	Map<String, Object> sourceScope
-	Map<String, Map<String, Object>> sourceTraitMaps
 	String sourceFieldName
-	String sourceFieldRpmType
-	List<String> typesToAdd
 
 	VisualizerRelInfo() {}
 
@@ -48,32 +40,19 @@ class VisualizerRelInfo
 		targetLevel = Long.valueOf(node.level as String)
 		targetScope = node.scope as CaseInsensitiveMap
 		scope = node.availableScope as CaseInsensitiveMap
-		loadTraits = node.loadTraits as boolean
-		typesToAdd = node.typesToAdd as List
-		hasFields = node.hasFields as boolean
 		group = setGroupName(allGroupsKeys)
 	}
 
 	Set<String> getRequiredScope()
 	{
 		Set<String> requiredScope = targetCube.getRequiredScope(targetScope, [:] as Map)
-		requiredScope.remove(AXIS_FIELD)
-		requiredScope.remove(AXIS_NAME)
-		requiredScope.remove(AXIS_TRAIT)
 		return requiredScope
 	}
 
 	String getDetails()
 	{
-		String effectiveName = effectiveNameByCubeName
 		StringBuilder sb = new StringBuilder()
 		String notesLabel = "<b>Note: </b>"
-
-		if (!hasFields)
-		{
-			sb.append("<b>*** Unable to load fields and traits for ${effectiveName}</b>${DOUBLE_BREAK}")
-			notesLabel = "<b>Reason: </b>"
-		}
 
 		//Notes
 		if (notes)
@@ -86,23 +65,12 @@ class VisualizerRelInfo
 		}
 
 		//Scope
-		if (hasFields)
-		{
-			if (loadTraits)
-			{
-				sb.append("<b>Utilized scope</b>")
-			}
-			else
-			{
-				sb.append("<b>Utilized scope to load class without all traits</b>")
-
-			}
-			sb.append("<pre><ul>")
-			targetScope.each { String key, Object value ->
-				sb.append("<li>${key}: ${value}</li>")
-			}
-			sb.append("</ul></pre>${BREAK}")
+		sb.append("<b>Scope</b>")
+		sb.append("<pre><ul>")
+		targetScope.each { String key, Object value ->
+			sb.append("<li>${key}: ${value}</li>")
 		}
+		sb.append("</ul></pre>${BREAK}")
 
 		sb.append("<b>Available scope</b>")
 		sb.append("<pre><ul>")
@@ -111,72 +79,7 @@ class VisualizerRelInfo
 		}
 		sb.append("</ul></pre>${BREAK}")
 
-		//Fields
-		if (hasFields)
-		{
-			if (loadTraits)
-			{
-				addClassTraits(sb)
-			}
-			addFieldDetails(sb)
-		}
 		return sb.toString()
-	}
-
-	private void addFieldDetails(StringBuilder sb)
-	{
-		if (loadTraits)
-		{
-			sb.append("<b>Fields and traits</b>")
-	}
-		else
-		{
-			sb.append("<b>Fields</b>")
-		}
-		sb.append("<pre><ul>")
-		targetTraitMaps.each { String fieldName, v ->
-			if (CLASS_TRAITS != fieldName)
-			{
-				if (loadTraits)
-				{
-					sb.append("<li><b>${fieldName}</b></li>")
-					addTraits(sb, fieldName)
-				}
-				else
-				{
-					sb.append("<li>${fieldName}</li>")
-				}
-			}
-		}
-		sb.append("</ul></pre>")
-	}
-
-	private void addTraits(StringBuilder sb, String fieldName)
-	{
-		Map<String, Object> traits = targetTraitMaps[fieldName].sort() as Map
-		sb.append("<pre><ul>")
-		traits.each { String traitName, Object traitValue ->
-			if (traitValue != null)
-			{
-				String traitString = traitValue.toString()
-				if (traitString.startsWith(HTTP) || traitString.startsWith(HTTPS) || traitString.startsWith(FILE))
-				{
-					sb.append("<li>${traitName}: <a href=\"${traitString}\" target=\"_blank\">${traitString}</a></li>")
-				}
-				else
-				{
-					sb.append("<li>${traitName}: ${traitValue}</li>")
-				}
-			}
-		}
-		sb.append("</ul></pre>")
-	}
-
-	private void addClassTraits(StringBuilder sb)
-	{
-		sb.append("<b>Class traits</b>")
-		addTraits(sb, CLASS_TRAITS)
-		sb.append("${BREAK}")
 	}
 
 	String getNodeGroup(String groupSuffix)
@@ -185,118 +88,50 @@ class VisualizerRelInfo
 		{
 			throw new IllegalStateException("Group must be set prior to getting node group.")
 		}
-
-		return targetCube.name.startsWith(RPM_ENUM) ? group + groupSuffix : group
+		return group
 	}
 
 	String setGroupName(Set<String> allGroupsKeys, String cubeName = targetCube.name)
 	{
-		Iterable<String> splits = Splitter.on('.').split(cubeName)
-		String group = splits[2].toUpperCase()
-		this.group = allGroupsKeys.contains(group) ? group : UNSPECIFIED
+		this.group = UNSPECIFIED
 	}
 
-	String getSourceEffectiveName()
-	{
-		String scopedName = sourceScopedName
-		return scopedName == null ? sourceCube.name : scopedName
-	}
-
-	String getTargetEffectiveName()
-	{
-		String scopedName = targetScopedName
-		return scopedName == null ? targetCube.name : scopedName
-	}
-
-	String getSourceScopedName()
-	{
-		Map<String, Object> classTraitsTraitMap = sourceTraitMaps[CLASS_TRAITS] as Map
-		return classTraitsTraitMap ? classTraitsTraitMap[R_SCOPED_NAME] : null
-	}
-
-	String getTargetScopedName()
-	{
-		Map<String, Object> classTraitsTraitMap = targetTraitMaps[CLASS_TRAITS] as Map
-		return classTraitsTraitMap ? classTraitsTraitMap[R_SCOPED_NAME] : null
-	}
-
-	private static String getDotSuffix(String value)
+	protected static String getDotSuffix(String value)
 	{
 		int lastIndexOfDot = value.lastIndexOf('.')
 		return lastIndexOfDot == -1 ? value : value.substring(lastIndexOfDot + 1)
 	}
 
-	private static String getDotPrefix(String value) {
+	protected static String getDotPrefix(String value) {
 		int indexOfDot = value.indexOf('.')
 		return indexOfDot == -1 ? value : value.substring(0, value.indexOf('.'))
 	}
 
+	String getSourceEffectiveName()
+	{
+		return sourceCube.name
+	}
+
+	String getTargetEffectiveName()
+	{
+		return targetCube.name
+	}
+
 	String getNextTargetCubeName(String targetFieldName)
 	{
-		if (sourceCube.getAxis(AXIS_TRAIT).findColumn(R_SCOPED_NAME))
-		{
-			String scopedName = sourceTraitMaps[CLASS_TRAITS][R_SCOPED_NAME]
-			return !scopedName ?: RPM_CLASS_DOT + sourceFieldRpmType
-		}
-		return RPM_CLASS_DOT + targetFieldName
+		return null  //TODO
 	}
 
     String getSourceMessage()
     {
-        if (sourceTraitMaps)
-        {
-            return sourceScopedName ? ", the target of ${sourceScopedName} on ${getCubeDisplayName(sourceCube.name)}" : ""
-        }
-        return ''
+       return ''  //TODO
     }
 
-    void getTraitMaps(VisualizerHelper helper, ApplicationID appId, VisualizerInfo visInfo)
-    {
-        targetTraitMaps = [:]
-        Map output = [:]
-        if (targetCube.name.startsWith(RPM_ENUM))
-        {
-            helper.loadRpmClassFields(appId, RPM_ENUM, targetCube.name - RPM_ENUM_DOT, scope, targetTraitMaps, loadTraits, output)
-        }
-        else
-        {
-            helper.loadRpmClassFields(appId, RPM_CLASS, targetCube.name - RPM_CLASS_DOT, scope, targetTraitMaps, loadTraits, output)
-        }
-        removeNotExistsFields()
-		addRequiredAndOptionalScopeKeys(visInfo)
-        retainUsedScope(visInfo, output)
-    }
-
-    void retainUsedScope(VisualizerInfo visInfo, Map output)
-    {
-        Set<String> scopeCollector = new CaseInsensitiveSet<>()
-        scopeCollector.addAll(visInfo.requiredScopeKeys[targetCube.name])
-        scopeCollector.addAll(visInfo.optionalScopeKeys[targetCube.name])
-        scopeCollector << EFFECTIVE_VERSION_SCOPE_KEY
-
-        RuleInfo ruleInfo = NCube.getRuleInfo(output)
-        Set keysUsed = ruleInfo.getInputKeysUsed()
-        scopeCollector.addAll(keysUsed)
-
-        targetScope = new CaseInsensitiveMap(scope)
-        cullScope(targetScope.keySet(), scopeCollector)
-    }
-
-    void removeNotExistsFields()
-    {
-        targetTraitMaps.keySet().removeAll { String fieldName -> !targetTraitMaps[fieldName][R_EXISTS] }
-    }
-
-    static void cullScope(Set<String> scopeKeys, Set scopeCollector)
-    {
-        scopeKeys.removeAll { String item -> !(scopeCollector.contains(item) || item.startsWith(SYSTEM_SCOPE_KEY_PREFIX)) }
-    }
-
-	/**
+  	/**
 	 *  If the required and optional scope keys have not already been loaded for this cube,
 	 *  load them.
 	 */
-	private void addRequiredAndOptionalScopeKeys(VisualizerInfo visInfo)
+	protected void addRequiredAndOptionalScopeKeys(VisualizerInfo visInfo)
 	{
 		String cubeName = targetCube.name
 		if (!visInfo.requiredScopeKeys.containsKey(cubeName))
@@ -309,31 +144,15 @@ class VisualizerRelInfo
 	Map<String, Object> createEdge(int edgeId)
 	{
 		String sourceFieldName = sourceFieldName
-		Map<String, Map<String, Object>> sourceTraitMaps = sourceTraitMaps
-
 		Map<String, Object> edge = [:]
-		String sourceCubeEffectiveName = sourceEffectiveName
-		String targetCubeEffectiveName = targetEffectiveName
 		edge.id = String.valueOf(edgeId + 1)
 		edge.from = String.valueOf(sourceId)
 		edge.to = String.valueOf(targetId)
-		edge.fromName = sourceCubeEffectiveName
-		edge.toName = targetCubeEffectiveName
+		edge.fromName = sourceEffectiveName
+		edge.toName = targetEffectiveName
 		edge.fromFieldName = sourceFieldName
 		edge.level = String.valueOf(targetLevel)
-		Map<String, Map<String, Object>> sourceFieldTraitMap = sourceTraitMaps[sourceFieldName] as Map
-		String vMin = sourceFieldTraitMap[V_MIN] as String ?: '0'
-		String vMax = sourceFieldTraitMap[V_MAX] as String ?: '999999'
-
-		if (targetCube.name.startsWith(RPM_ENUM_DOT))
-		{
-			edge.label = sourceFieldName
-			edge.title = "Field ${sourceFieldName} cardinality ${vMin}:${vMax}".toString()
-		}
-		else
-		{
-			edge.title = "Valid value ${sourceFieldName} cardinality ${vMin}:${vMax}".toString()
-		}
+		edge.title = sourceFieldName  //TODO
 
 		return edge
 	}
@@ -356,94 +175,44 @@ class VisualizerRelInfo
 		node.sourceDescription = sourceCubeName ? sourceDescription : null
 		String detailsTitle1 = cubeDetailsTitle1
 
-		if (targetCubeName.startsWith(RPM_CLASS_DOT))
-		{
-			node.label = getDotSuffix(targetEffectiveName)
-			node.detailsTitle1 = detailsTitle1
-			node.detailsTitle2 = cubeDetailsTitle2
-			node.title = getCubeDisplayName(targetCubeName)
-			node.typesToAdd = typesToAdd
-		}
-		else
-		{
-			node.detailsTitle1 = detailsTitle1
-			node.title = detailsTitle1
-		}
+		node.label = nodeLabel
+		node.detailsTitle1 = detailsTitle1
+		node.detailsTitle2 = cubeDetailsTitle2
+		node.title = getCubeDisplayName(targetCubeName)
+
 		node.details = details
 		group ?: setGroupName(allGroupsKeys)
 		node.group = getNodeGroup(groupSuffix)
-		node.loadTraits = loadTraits
-		node.hasFields = hasFields
 		return node
+	}
+
+	protected String getNodeLabel()
+	{
+		targetEffectiveName
 	}
 
 	String getEffectiveNameByCubeName()
 	{
-		String scopeKey = getDotSuffix(targetCube.name)
-		return scope[scopeKey] ?: targetEffectiveName
+		return targetCube.name
 	}
 
-	static String getCubeDisplayName(String cubeName)
+	String getCubeDisplayName(String cubeName)
 	{
-		String displayName
-		if (cubeName.startsWith(RPM_CLASS_DOT))
-		{
-			displayName = cubeName - RPM_CLASS_DOT
-		}
-		else if (cubeName.startsWith(RPM_ENUM_DOT))
-		{
-			displayName = cubeName - RPM_ENUM_DOT
-		}
-		return displayName
+		return cubeName
 	}
 
 	String getSourceDescription()
 	{
-		String description
-		String sourceCubeName = sourceCube.name
-		if (sourceCubeName.startsWith(RPM_CLASS_DOT))
-		{
-			description = getDotSuffix(sourceEffectiveName)
-		}
-		else if (sourceCubeName.startsWith(RPM_ENUM_DOT))
-		{
-			if (targetScopedName)
-			{
-				String sourceDisplayName = getCubeDisplayName(sourceCubeName)
-				String scopeKeyForSourceOfSource = getDotPrefix(sourceDisplayName)
-				String nameOfSourceOfSource = sourceScope[scopeKeyForSourceOfSource]
-				String fieldNameSourceOfSource = sourceScope[SOURCE_FIELD_NAME]
-				description = "field ${fieldNameSourceOfSource} on ${nameOfSourceOfSource}".toString()
-			}
-			else{
-				description = getCubeDisplayName(sourceCubeName)
-			}
-		}
-		return description
+		return sourceCube.name
 	}
 
 	String getCubeDetailsTitle1()
 	{
-		String detailsTitle
-		String targetCubeName = targetCube.name
-		if (targetCubeName.startsWith(RPM_CLASS_DOT))
-		{
-			detailsTitle = getCubeDisplayName(targetCubeName)
-		}
-		else if (targetCubeName.startsWith(RPM_ENUM_DOT))
-		{
-			String sourceName =  sourceTraitMaps ? getDotSuffix(sourceEffectiveName) : getCubeDisplayName(sourceCube.name)
-			detailsTitle = "Valid values for field ${sourceFieldName} on ${sourceName}".toString()
-		}
-		return detailsTitle
+		return targetCube.name
 	}
 
 	String getCubeDetailsTitle2()
 	{
-		if (targetCube.name.startsWith(RPM_CLASS_DOT) && getTargetScopedName())
-		{
-			return getEffectiveNameByCubeName()
-		}
 		return null
 	}
 }
