+++ src/main/webapp/html/visualize.html
@@ -36,8 +36,8 @@
                         <label>Scope: </label>
                     </div>
                     <div id="scope-div" class="inline-block">
-                        <div class="input-group input-group-sm col-width-large" title="Scope used to load graph data from n-cube class definitions">
-                            <input id="scope" type="text" placeholder="Loading scope..." >
+                        <div class="input-group input-group-sm col-width-large" title="Scope used to load graph data from n-cube definitions">
+                            <input id="scope" type="text" placeholder="Add scope..." >
                         </div>
                     </div>
                 </div>
@@ -166,7 +166,7 @@
     <div id="visLayoutEast" class="pane ui-layout-east" aria-disabled="false">
         <h3 id="nodeDetailsTitle1"></h3>
         <h3 id="nodeDetailsTitle2"></h3>
-        <div id="nodeVisualizer" title="Link to new visualization with this class as the starting point"></div>
+        <div id="nodeVisualizer" title="Link to new visualization with this node as the starting point"></div>
         <div id="nodeTraits" title="Show or hide traits for the class"></div>
         <div id="nodeCubeLink" title="Link to the n-cube view"></div>
         <div id="nodeAddTypes" title="Add to this class" class="dropdown"></div>
